CSM fixes and distance travel fixes
Cascaded shadow mapping errors solved (Shadows dissapearing). Error in
ortho matrix generation solved.
View is now totally centered in player (Distance visual bugs solved).
Terrain generation for big distances / collisions in far distances
fixed. FloatBuffer possible depletion fixed, always recycling now. Chunk
loading bugs in far distances fixed.
World boundaries now are in the int value range (From Integer.MIN_VALUE
to Integer.MAX_VALUE meters). That gives us ~4 million km length.

+++ src/monecruft/utils/VoxelUtils.java
@@ -3,7 +3,7 @@ package monecruft.utils;
 public class VoxelUtils 
 {
 	
-	public static int trueMod(float val,int toMod)
+	public static int trueMod(double val,int toMod)
 	{
 		if(val>0)
 		{
